backend/qa/requirements.txt
# Vector database
chromadb==1.3.4  # ChromaDB version 1.3.4

# Web service
flask
flask-cors  # Enable CORS for frontend communication
python-dotenv

# LLM providers (choose one or more)
google-genai  # For gemini-2.5-flash and newer models
google-generativeai  # Legacy support (if needed)

# for reranking
sentence_transformers